requirements.txt
asarPy==1.0.1
blinker==1.9.0
click==8.3.0
Cython==3.1.6
Flask==3.1.2
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
gunicorn==21.2.0
psycopg2-binary==2.9.7
Markdown==3.9
MarkupSafe==3.0.3
meson==1.9.1
mysql-connector-python==9.5.0
packaging==25.0
PyMySQL==1.1.2
setuptools==80.9.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
wheel==0.46.1
